thumbnail issue: cannot read property [0] fixed for the DLCS content
