+++ src/collections/test.json
@@ -301,10 +301,10 @@
       },
       "thumbnail": [
         { 
-          "id": "https://dlc.services/thumbs/5/1/dbc3f644-5830-4d71-87c9-db12e35d1964/full/full/0/default.jpg", 
+          "id": "https://dlc.services/thumbs/5/1/2378e17e-f976-419f-92f7-cb3bb883e009/full/full/0/default.jpg", 
           "service": {
             "@context":"http://iiif.io/api/image/2/context.json",
-            "@id":"https://dlc.services/thumbs/5/1/dbc3f644-5830-4d71-87c9-db12e35d1964",
+            "@id":"https://dlc.services/thumbs/5/1/2378e17e-f976-419f-92f7-cb3bb883e009",
             "profile": [
               "http://iiif.io/api/image/2/level0.json",
               {
