+++ src/objects/multi-canvas-object.json
@@ -796,5 +796,5 @@
       "id": "https://delft-static-site-generator.netlify.com/iiif/18fe5949-601f-b19f-1cde-f78f8f6caa28/canvas/95204d48-eed3-8a29-dff1-fc7eda57a09a"
     }
   ],
-  "id": "https://delft-static-site-generator.netlify.com/iiif/18fe5949-601f-b19f-1cde-f78f8f6caa28/manifest"
+  "id": "https://delft-static-site-generator.netlify.com/objects/multi-canvas-object"
 }
