+++ src/objects/object-1.json
@@ -2,6 +2,9 @@
   "label": {
     "en": [
       "First test object"
+    ],
+    "nl": [
+      "[NL]First test object"
     ]
   },
   "type": "Manifest",
@@ -11,6 +14,9 @@
       "label": {
         "en": [
           "Canvas 1"
+        ],
+        "nl": [
+          "[NL]Canvas 1"
         ]
       },
       "height": 1669,
@@ -28,6 +34,9 @@
               "label": {
                 "en": [
                   "Image Annotation 1"
+                ],
+                "nl": [
+                  "[NL]Image Annotation 1"
                 ]
               },
               "body": {
@@ -36,7 +45,7 @@
                 "format": "image/jpg",
                 "service": [
                   {
-                    "id": "https://dlc.services/iiif-img/5/1/2378e17e-f976-419f-92f7-Cb3bb883e009",
+                    "id": "https://dlc.services/iiif-img/5/1/2378e17e-f976-419f-92f7-cb3bb883e009",
                     "type": "ImageService3",
                     "profile": "level2"
                   }
